Export `ConfigSchema` type
In light of https://github.com/joe-bell/cva/discussions/146, I want to extract the config to a variable to be able to use in documentation (i.e. Storybook). Until https://github.com/joe-bell/cva/pull/33 is merged, there's no other way to reuse the variants. However, by doing so we lose the type safety of the config. I'd like to be able to verify that the variants config satisfies the expected config schema, and by exporting `ConfigSchema` it'll be as easy as `satisfies ConfigSchema`!